#P2PW-8 #P2PW-7 Closing issue saved, Crash fixed

+++ domain/src/main/java/com/wowlet/domain/usecases/PinCodeUseCase.kt
@@ -13,9 +13,9 @@ class PinCodeUseCase(private val preferenceService: PreferenceService) : PinCode
     override suspend fun initPinCode(pinCode: String): Boolean {
         val value = preferenceService.getPinCodeValue()
         return value?.let {
-            it.pinCode == pinCode
+            it.pinCode == pinCode.toInt()
         } ?: kotlin.run {
-            preferenceService.setPinCodeValue(PinCodeData(pinCode))
+            preferenceService.setPinCodeValue(PinCodeData(pinCode.toInt()))
             true
         }
     }
